Move research API to .cpp file
